MINIFICPP-1443 Fix 'possible loss of data' warnings
... and a few others which were related or nearby in the code.

+++ extensions/http-curl/client/HTTPClient.cpp
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 #include "HTTPClient.h"
-#include "Exception.h"
+
 #include <memory>
 #include <climits>
 #include <cinttypes>
@@ -24,6 +24,9 @@
 #include <vector>
 #include <string>
 #include <algorithm>
+
+#include "Exception.h"
+#include "utils/gsl.h"
 #include "utils/StringUtils.h"
 #include "utils/RegexUtils.h"
 
@@ -213,7 +216,7 @@ void HTTPClient::setContentType(std::string content_type) {
 }
 
 std::string HTTPClient::escape(std::string string_to_escape) {
-  return curl_easy_escape(http_session_, string_to_escape.c_str(), string_to_escape.length());
+  return curl_easy_escape(http_session_, string_to_escape.c_str(), gsl::narrow<int>(string_to_escape.length()));
 }
 
 void HTTPClient::setPostFields(const std::string& input) {
